requirements.txt
autopep8==1.5.4
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.10
django-bootstrap-form==3.4
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-phonenumber-field==4.0.0
django-timezone-field==4.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
jsonfield==1.0.3
Markdown==3.2.2
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.2.0
pycodestyle==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
urllib3==1.25.10
whitenoise==5.2.0
wincertstore==0.2
